ps: prevent situations where users were trying to circumvent the payment process

+++ parkstay/views.py
@@ -1,4 +1,4 @@
-
+import logging
 from django.db.models import Q
 from django.http import Http404, HttpResponse, JsonResponse
 from django.shortcuts import render, get_object_or_404, redirect
@@ -29,12 +29,15 @@ from parkstay.models import (Campground,
                                 )
 from parkstay import emails
 from ledger.accounts.models import EmailUser, Address
+from ledger.payments.models import Invoice
 from django_ical.views import ICalFeed
 from datetime import datetime, timedelta
 from decimal import *
 
 from parkstay.helpers import is_officer
 from parkstay import utils
+
+logger = logging.getLogger('booking_checkout')
 
 class CampsiteBookingSelector(TemplateView):
     template_name = 'ps/campsite_booking_selector.html'
@@ -311,22 +314,38 @@ class BookingSuccessView(TemplateView):
             booking = utils.get_session_booking(request.session)
             invoice_ref = request.GET.get('invoice')
 
-            # FIXME: replace with server side notify_url callback
-            book_inv, created = BookingInvoice.objects.get_or_create(booking=booking, invoice_reference=invoice_ref)
-            
-            # set booking to be permanent fixture
-            booking.booking_type = 1  # internet booking
-            booking.expiry_time = None
-            booking.save()
+            try:
+                inv = Invoice.objects.get(reference=invoice_ref)
+            except Invoice.DoesNotExist:
+                logger.error('User {} with id {} tried making a booking with an incorrect invoice'.format(booking.customer.get_full_name(),booking.customer.id))
+                return redirect('public_make_booking')
 
-            utils.delete_session_booking(request.session)
-            request.session['ps_last_booking'] = booking.id
-            
-            # send out the invoice before the confirmation is sent
-            emails.send_booking_invoice(booking)
-            # for fully paid bookings, fire off confirmation email
-            if booking.paid:
-                emails.send_booking_confirmation(booking,request)
+            if inv.system not in ['0019']:
+                logger.error('User {} with id {} tried making a booking with an invoice from another system with reference number {}'.format(booking.customer.get_full_name(),booking.customer.id,inv.reference))
+                return redirect('public_make_booking')
+
+            try:
+                b = BookingInvoice.objects.get(invoice_reference=invoice_ref)
+                logger.error('User {} with id {} tried making a booking with an already used invoice with reference number {}'.format(booking.customer.get_full_name(),booking.customer.id,inv.reference))
+                return redirect('public_make_booking')
+            except BookingInvoice.DoesNotExist:
+
+                # FIXME: replace with server side notify_url callback
+                book_inv, created = BookingInvoice.objects.get_or_create(booking=booking, invoice_reference=invoice_ref)
+
+                # set booking to be permanent fixture
+                booking.booking_type = 1  # internet booking
+                booking.expiry_time = None
+                booking.save()
+
+                utils.delete_session_booking(request.session)
+                request.session['ps_last_booking'] = booking.id
+
+                # send out the invoice before the confirmation is sent
+                emails.send_booking_invoice(booking)
+                # for fully paid bookings, fire off confirmation email
+                if booking.paid:
+                    emails.send_booking_confirmation(booking,request) 
 
         except Exception as e:
             if ('ps_last_booking' in request.session) and Booking.objects.filter(id=request.session['ps_last_booking']).exists():
